requirements.txt
# PubMed2Zhihu 依赖清单

# PubMed API
biopython>=1.81

# HTTP请求和解析
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web框架
Flask>=2.3.0
Jinja2>=3.1.0

# 工具库
colorama>=0.4.6
configparser>=6.0.0

# 浏览器自动化（用于截图获取图片）
playwright>=1.40.0

# PDF处理
PyMuPDF>=1.23.0

